test: repair hps tests
