--- tests/hps/conftest.py
@@ -1,17 +1,0 @@
-from pathlib import Path
-
-import pytest
-from _pytest.tmpdir import TempdirFactory
-
-
-@pytest.fixture(scope="module")
-def tmp_path_module(tmpdir_factory: TempdirFactory) -> Path:
-    path = tmpdir_factory.mktemp(__name__)
-    return Path(path)
-
-
-@pytest.fixture(scope="module")
-def retrain_tmp_path(tmp_path_module: Path) -> Path:
-    path = tmp_path_module.joinpath("retrain")
-    path.mkdir()
-    return path
