+++ tests/hps/test_hps_classification.py
@@ -6,6 +6,7 @@ from typing import Any, Dict, List, Tuple, Union
 import pandas as pd
 import pytest
 import yaml
+from _pytest.tmpdir import TempdirFactory
 
 from embeddings.hyperparameter_search.lighting_configspace import (
     LightingTextClassificationConfigSpace,
@@ -31,6 +32,19 @@ def load_dataset_kwargs() -> Dict[str, Union[List[str], str]]:
         "test_domains": ["hotels", "medicine"],
         "text_cfg": "text",
     }
+
+
+@pytest.fixture(scope="module")
+def tmp_path_module(tmpdir_factory: TempdirFactory) -> Path:
+    path = tmpdir_factory.mktemp(__name__)
+    return Path(path)
+
+
+@pytest.fixture(scope="module")
+def retrain_tmp_path(tmp_path_module: Path) -> Path:
+    path = tmp_path_module.joinpath("retrain")
+    path.mkdir()
+    return path
 
 
 @pytest.fixture(scope="module")
@@ -193,7 +207,7 @@ def test_hparams_best_params_files_compatibility(
 ) -> None:
     df, metadata = classification_hps_run_result
 
-    with open(retrain_tmp_path / "lightning_logs" / "version_0" / "hparams.yaml") as f:
+    with open(retrain_tmp_path / "csv" / "version_0" / "hparams.yaml") as f:
         hparams = yaml.load(f, Loader=yaml.Loader)
 
     with open(tmp_path_module / "best_params.yaml") as f:
